+++ packages/terra-form-textarea/src/TextareaField.jsx
@@ -72,6 +72,10 @@ const propTypes = {
    */
   onChange: PropTypes.func,
   /**
+   * Placeholder text.
+   */
+  placeholder: PropTypes.string,
+  /**
    * Whether or not the field is required.
    */
   required: PropTypes.bool,
@@ -102,6 +106,7 @@ const defaultProps = {
   labelAttrs: {},
   maxWidth: undefined,
   onChange: undefined,
+  placeholder: undefined,
   required: false,
   showOptional: false,
   value: undefined,
@@ -126,6 +131,7 @@ const TextareaField = (props) => {
     required,
     showOptional,
     onChange,
+    placeholder,
     value,
     ...customProps
   } = props;
@@ -166,6 +172,7 @@ const TextareaField = (props) => {
         disabled={inputAttrs.disabled || disabled}
         id={inputId}
         onChange={onChange}
+        placeholder={placeholder || inputAttrs.placeholder}
         value={value}
         defaultValue={defaultValue}
         aria-describedby={ariaDescriptionIds}