+++ packages/terra-form-textarea/CHANGELOG.md
@@ -6,6 +6,9 @@ Unreleased
 ### Breaking Changes
 * `label` prop's type changed from node to string.
 
+### Added
+* Added `placeholder` as a property to `<Textarea />` and `<TextareaField />`
+
 3.28.0 - (October 3, 2019)
 ------------------
 ### Changed